feat: rework roles/permissions systems + add crud for roles

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,7 +13,8 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
+            $table->string('first_name')->nullable();
+            $table->string('last_name')->nullable();
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
@@ -27,6 +28,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
+        if ( ! app()->isProduction()) {
+            Schema::dropIfExists('users');
+        }
     }
 };
