+++ database/migrations/2023_09_03_134842_create_permission_tables.php
@@ -135,10 +135,12 @@ class CreatePermissionTables extends Migration
             throw new \Exception('Error: config/permission.php not found and defaults could not be merged. Please publish the package configuration before proceeding, or drop the tables manually.');
         }
 
-        Schema::drop($tableNames['role_has_permissions']);
-        Schema::drop($tableNames['model_has_roles']);
-        Schema::drop($tableNames['model_has_permissions']);
-        Schema::drop($tableNames['roles']);
-        Schema::drop($tableNames['permissions']);
+        if ( ! app()->isProduction()) {
+            Schema::drop($tableNames['role_has_permissions']);
+            Schema::drop($tableNames['model_has_roles']);
+            Schema::drop($tableNames['model_has_permissions']);
+            Schema::drop($tableNames['roles']);
+            Schema::drop($tableNames['permissions']);
+        }
     }
 }
